One plugin one action changes for shell on the agent
+ Adds a shell plugin and an IShellAction abstraction to conform to same
pattern we use for kube/web/db. Refactors existing shell plugin to be
the UnixShell action which is currently the only action in the shell
plugin

+ Also simplify ShellInputMessage by using []byte instead of string so
json marshing automaticaly encode/decodes as base64 string

+ Refactors shell/datachanell unit tests to work with these changes

+ Fix unit tests now tht
